upload code edit lichlamviec user, donvi, delete lichlamviec user, donvi

+++ src/main/java/com/example/qllichlamviec/controller/LichLamViecController.java
@@ -1,29 +1,19 @@
 package com.example.qllichlamviec.controller;
 
-import com.example.qllichlamviec.modal.dto.LichLanViecDTO;
-import com.example.qllichlamviec.modal.dto.TaiKhoanDangNhapDTO;
-import com.example.qllichlamviec.modal.dto.TaiKhoanNguoiDungDTO;
-import com.example.qllichlamviec.modal.dto.ThongBaoDTO;
+import com.example.qllichlamviec.modal.dto.*;
 import com.example.qllichlamviec.modal.system.Error;
-import com.example.qllichlamviec.modal.system.NguoiDungDangNhapDTO;
 import com.example.qllichlamviec.service.*;
 import com.example.qllichlamviec.util.*;
-import org.apache.catalina.security.SecurityUtil;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.modelmapper.ModelMapper;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.validation.Valid;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Pattern;
 
@@ -112,7 +102,7 @@ public class LichLamViecController {
 //    -----------------------
 
     @PostMapping("/them")
-    public ResponseEntity<Object> themLichLamViec(@RequestBody LichLanViecDTO lichLamViecDTO, HttpServletRequest httpRequest) {
+    public ResponseEntity<Object> themLichLamViec(@RequestBody LichLamViecDTO lichLamViecDTO, HttpServletRequest httpRequest) {
         try {
 
             TaiKhoan taiKhoan = taiKhoanService.getTaiKhoanFromRequest(httpRequest);
@@ -158,9 +148,41 @@ public class LichLamViecController {
         }
 
     }
+
+    @PutMapping("/edit-lich-donvi")
+    public ResponseEntity<Object> EditLichDonVi(@RequestBody LichLamViecDTO lichLamViecDTO, @RequestParam String idLich, HttpServletRequest httpRequest) {
+        try {
+
+            TaiKhoan taiKhoan = taiKhoanService.getTaiKhoanFromRequest(httpRequest);
+            if (taiKhoan == null) {
+                return new ResponseEntity<>("Không tìm thấy thông tin người dùng", HttpStatus.UNAUTHORIZED);
+            }
+
+            LichLamViec lichLamViec = lichLamViecService.getById(idLich);
+            if (lichLamViec == null) {
+                return new ResponseEntity<>("Không tìm thấy lịch làm việc", HttpStatus.NOT_FOUND);
+            }
+
+            lichLamViec.setThoiGianBD(lichLamViecDTO.getThoiGianBD());
+            lichLamViec.setThoiGianKT(lichLamViecDTO.getThoiGianKT());
+            lichLamViec.setDiaDiem(lichLamViecDTO.getDiaDiem());
+            lichLamViec.setNoiDung(lichLamViecDTO.getNoiDung());
+            lichLamViec.setTieuDe(lichLamViecDTO.getTieuDe());
+            lichLamViec.setGhiChu(lichLamViecDTO.getGhiChu());
+
+            LichLamViec lichLamViecDaTao = lichLamViecService.save(lichLamViec);
+
+            return new ResponseEntity<>(lichLamViecDaTao, HttpStatus.CREATED);
+
+        } catch (Exception e) {
+            return new ResponseEntity<>("Lỗi khi sửa lịch làm việc: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+    }
+
     @PreAuthorize("hasRole('ROLE_USER') or hasRole('ROLE_ADMIN')")
     @PostMapping("/them-lich-user")
-    public ResponseEntity<Object> themLicUser(@RequestBody LichLanViecDTO lichLamViecDTO, HttpServletRequest httpRequest) {
+    public ResponseEntity<Object> themLichUser(@RequestBody LichCaNhanDTO lichCaNhanDTO, HttpServletRequest httpRequest) {
         try {
 
             TaiKhoan taiKhoan = taiKhoanService.getTaiKhoanFromRequest(httpRequest);
@@ -170,12 +192,12 @@ public class LichLamViecController {
             }
 
             LichLamViec lichLamViec = new LichLamViec();
-            lichLamViec.setThoiGianBD(lichLamViecDTO.getThoiGianBD());
-            lichLamViec.setThoiGianKT(lichLamViecDTO.getThoiGianKT());
-            lichLamViec.setDiaDiem(lichLamViecDTO.getDiaDiem());
-            lichLamViec.setNoiDung(lichLamViecDTO.getNoiDung());
-            lichLamViec.setTieuDe(lichLamViecDTO.getTieuDe());
-            lichLamViec.setGhiChu(lichLamViecDTO.getGhiChu());
+            lichLamViec.setThoiGianBD(lichCaNhanDTO.getThoiGianBD());
+            lichLamViec.setThoiGianKT(lichCaNhanDTO.getThoiGianKT());
+            lichLamViec.setDiaDiem(lichCaNhanDTO.getDiaDiem());
+            lichLamViec.setNoiDung(lichCaNhanDTO.getNoiDung());
+            lichLamViec.setTieuDe(lichCaNhanDTO.getTieuDe());
+            lichLamViec.setGhiChu(lichCaNhanDTO.getGhiChu());
 
             lichLamViec.setTaiKhoan(taiKhoan);
 
@@ -189,14 +211,58 @@ public class LichLamViecController {
 
     }
 
+    @PutMapping("/edit-lich-user")
+    public ResponseEntity<Object> EditLichUser(@RequestBody LichCaNhanDTO lichCaNhanDTO, @RequestParam String idLich, HttpServletRequest httpRequest) {
+        try {
+
+            TaiKhoan taiKhoan = taiKhoanService.getTaiKhoanFromRequest(httpRequest);
+            if (taiKhoan == null) {
+                return new ResponseEntity<>("Không tìm thấy thông tin người dùng", HttpStatus.UNAUTHORIZED);
+            }
+
+            LichLamViec lichLamViec = lichLamViecService.getById(idLich);
+            if (lichLamViec == null) {
+                return new ResponseEntity<>("Không tìm thấy lịch làm việc", HttpStatus.NOT_FOUND);
+            }
+
+            String tk = lichLamViec.getTaiKhoan().get_id().toHexString();
+            String tk2 = taiKhoan.get_id().toHexString();
+
+            if (tk2.equals(tk)) {
+                lichLamViec.setThoiGianBD(lichCaNhanDTO.getThoiGianBD());
+                lichLamViec.setThoiGianKT(lichCaNhanDTO.getThoiGianKT());
+                lichLamViec.setDiaDiem(lichCaNhanDTO.getDiaDiem());
+                lichLamViec.setNoiDung(lichCaNhanDTO.getNoiDung());
+                lichLamViec.setTieuDe(lichCaNhanDTO.getTieuDe());
+                lichLamViec.setGhiChu(lichCaNhanDTO.getGhiChu());
+
+                LichLamViec lichLamViecDaTao = lichLamViecService.save(lichLamViec);
+                return new ResponseEntity<>(lichLamViecDaTao, HttpStatus.CREATED);
+            }else {
+                return new ResponseEntity<>("Đây là lịch cá nhân bạn không có quyền chỉnh sữa", HttpStatus.BAD_REQUEST);
+            }
+
+        } catch (Exception e) {
+            return new ResponseEntity<>("Lỗi khi sửa lịch làm việc: " + e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+    }
+
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @DeleteMapping("/{id}")
     public ResponseEntity<Object>deleteByAdmin(@PathVariable String id, HttpServletRequest httpRequest){
         try {
-            lichLamViecService.deleteByID(id);
+            TaiKhoan tk = taiKhoanService.getTaiKhoanFromRequest(httpRequest);
+            LichLamViec lichLamViec = lichLamViecService.getById(id);
+            if (lichLamViec.getDonVi() != null || lichLamViec.getTaiKhoan().get_id().toString().equals(tk.get_id().toString())) {
+//                TaiKhoan kttk = taiKhoanService.kiemTraUserAdmin(();
+                lichLamViecService.deleteByID(id);
+            }
+
             return new ResponseEntity<>("Xoá thành công lịch làm việc với ID: "+id, HttpStatus.OK);
+
         }catch (Exception e){
-            return new ResponseEntity<>(new Error("400", e.getMessage()), HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new Error("400","Đây là lịch cá nhân bạn không có quyền xoa", e.getMessage()), HttpStatus.BAD_REQUEST);
         }
     }
 
@@ -223,9 +289,4 @@ public class LichLamViecController {
             return new ResponseEntity<>(new Error("400", e.getMessage()), HttpStatus.BAD_REQUEST);
         }
     }
-
-
-
-
-
 }
